requirements.txt
flask
scikit-learn
numpy
pytesseract
pdf2image
Pillow
easyocr
sentence-transformers
